Bella-Office-Orchestrator/requirements copy.txt
streamlit==1.35.0
python-dotenv
requests
slack_sdk

#preprocessing
psycopg2-binary
pandas
boto3            

openai

fastapi
uvicorn
slack-sdk

google-auth-oauthlib
google-api-python-client
requests
APScheduler

pandas
fpdf
httpx
pytz
sqlalchemy
